Add BlazeJavaElementFinder class
This CL adds `BlazeJavaElementFinder` class as a special extension point of `java.elementFinder`. Some classes can exist in the `META-INF/TRANSITIVE` folder of a library and at the same time exist in their actual library. When a class found in multiple files, [`JavaElementFinder`](https://github.com/JetBrains/kotlin/blob/master/compiler/light-classes/src/org/jetbrains/kotlin/asJava/finder/JavaElementFinder.kt) sorts the files and returns the first one which in some cases will not be the actual library of the class. This causes the navigation to the source of this class to not be possible besides other errors that will show because of this wrong reference resolution.

In `BlazeJavaElementFinder`, we get the classes found and sorted by `JavaElementFinder` then return the first location of the class file that is not in `META-INF/TRANSITIVE` if any.

PiperOrigin-RevId: 437130147