Use sync strategy to determine blaze info binary path
Adds a utility method to BuildSystem for returning a binary path. If the sync strategy is PARALLEL, the method returns the path from a parallel invoker, if it exists. In all other cases, the default path is returned.

This ensures that initial invocations of `info` and `query` use the parallel invoker binary when applicable. When sync strategy is DECIDE_AUTOMATICALLY, the local invoker is used through the query invocation.

PiperOrigin-RevId: 451024320